+++ deepforest/__init__.py
@@ -1,14 +1,19 @@
-from .cascade import CascadeForestClassifier
-from .forest import RandomForestClassifier
-from .forest import ExtraTreesClassifier
-from .tree import DecisionTreeClassifier
-from .tree import ExtraTreeClassifier
+from .cascade import CascadeForestClassifier, CascadeForestRegressor
+from .forest import RandomForestClassifier, RandomForestRegressor
+from .forest import ExtraTreesClassifier, ExtraTreesRegressor
+from .tree import DecisionTreeClassifier, DecisionTreeRegressor
+from .tree import ExtraTreeClassifier, ExtraTreeRegressor
 
 
 __all__ = [
     "CascadeForestClassifier",
+    "CascadeForestRegressor",
     "RandomForestClassifier",
+    "RandomForestRegressor",
     "ExtraTreesClassifier",
+    "ExtraTreesRegressor",
     "DecisionTreeClassifier",
+    "DecisionTreeRegressor",
     "ExtraTreeClassifier",
+    "ExtraTreeRegressor",
 ]