+++ deepforest/_estimator.py
@@ -3,10 +3,15 @@
 
 __all__ = ["Estimator"]
 
-from .forest import RandomForestClassifier, ExtraTreesClassifier
+from .forest import (
+    RandomForestClassifier,
+    ExtraTreesClassifier,
+    RandomForestRegressor,
+    ExtraTreesRegressor,
+)
 
 
-def make_estimator(
+def make_classifier_estimator(
     name,
     n_trees=100,
     max_depth=None,
@@ -39,6 +44,39 @@ def make_estimator(
     return estimator
 
 
+def make_regressor_estimator(
+    name,
+    n_trees=100,
+    max_depth=None,
+    min_samples_leaf=1,
+    n_jobs=None,
+    random_state=None,
+):
+    # RandomForestRegressor
+    if name == "rf":
+        estimator = RandomForestRegressor(
+            n_estimators=n_trees,
+            max_depth=max_depth,
+            min_samples_leaf=min_samples_leaf,
+            n_jobs=n_jobs,
+            random_state=random_state,
+        )
+    # ExtraTreesRegressor
+    elif name == "erf":
+        estimator = ExtraTreesRegressor(
+            n_estimators=n_trees,
+            max_depth=max_depth,
+            min_samples_leaf=min_samples_leaf,
+            n_jobs=n_jobs,
+            random_state=random_state,
+        )
+    else:
+        msg = "Unknown type of estimator, which should be one of {{rf, erf}}."
+        raise NotImplementedError(msg)
+
+    return estimator
+
+
 class Estimator(object):
     def __init__(
         self,
@@ -48,10 +86,28 @@ class Estimator(object):
         min_samples_leaf=1,
         n_jobs=None,
         random_state=None,
+        is_classifier=True,
     ):
-        self.estimator_ = make_estimator(
-            name, n_trees, max_depth, min_samples_leaf, n_jobs, random_state
-        )
+
+        self.is_classifier = is_classifier
+        if self.is_classifier:
+            self.estimator_ = make_classifier_estimator(
+                name,
+                n_trees,
+                max_depth,
+                min_samples_leaf,
+                n_jobs,
+                random_state,
+            )
+        else:
+            self.estimator_ = make_regressor_estimator(
+                name,
+                n_trees,
+                max_depth,
+                min_samples_leaf,
+                n_jobs,
+                random_state,
+            )
 
     @property
     def oob_decision_function_(self):
@@ -64,8 +120,11 @@ class Estimator(object):
         return X_aug
 
     def transform(self, X):
-
-        return self.estimator_.predict_proba(X)
+        if self.is_classifier:
+            return self.estimator_.predict_proba(X)
+        return self.estimator_.predict(X)
 
     def predict(self, X):
-        return self.estimator_.predict_proba(X)
+        if self.is_classifier:
+            return self.estimator_.predict_proba(X)
+        return self.estimator_.predict(X)