+++ deepforest/tree/__init__.py
@@ -1,6 +1,13 @@
 from .tree import BaseDecisionTree
 from .tree import DecisionTreeClassifier
+from .tree import DecisionTreeRegressor
 from .tree import ExtraTreeClassifier
+from .tree import ExtraTreeRegressor
 
-
-__all__ = ["BaseDecisionTree", "DecisionTreeClassifier", "ExtraTreeClassifier"]
+__all__ = [
+    "BaseDecisionTree",
+    "DecisionTreeClassifier",
+    "DecisionTreeRegressor",
+    "ExtraTreeClassifier",
+    "ExtraTreeRegressor",
+]