+++ tests/test_forest.py
@@ -2,11 +2,13 @@ import pytest
 
 from deepforest import RandomForestClassifier
 from deepforest import ExtraTreesClassifier
+from deepforest import RandomForestRegressor
+from deepforest import ExtraTreesRegressor
 from deepforest.forest import _get_n_samples_bootstrap
 
 # Load utils
 from sklearn.ensemble._hist_gradient_boosting.binning import _BinMapper
-from sklearn.datasets import load_iris, load_wine
+from sklearn.datasets import load_iris, load_wine, load_boston
 from sklearn.ensemble._forest import (
     _get_n_samples_bootstrap as sklearn_get_n_samples_bootstrap,
 )
@@ -51,7 +53,7 @@ def test_n_samples_bootstrap_invalid_type():
 
 
 @pytest.mark.parametrize("load_func", [load_iris, load_wine])
-def test_forest_workflow(load_func):
+def test_forest_classifier_workflow(load_func):
 
     n_estimators = 100  # to avoid oob warning
     random_state = 42
@@ -72,6 +74,35 @@ def test_forest_workflow(load_func):
 
     # Extremely Random Forest
     model = ExtraTreesClassifier(
+        n_estimators=n_estimators, random_state=random_state
+    )
+
+    model.fit(X_binned, y)
+    model.predict(X_binned)
+
+
+@pytest.mark.parametrize("load_func", [load_boston])
+def test_forest_regressor_workflow(load_func):
+
+    n_estimators = 100  # to avoid oob warning
+    random_state = 42
+
+    X, y = load_func(return_X_y=True)
+
+    # Data binning
+    binner = _BinMapper(random_state=random_state)
+    X_binned = binner.fit_transform(X)
+
+    # Random Forest
+    model = RandomForestRegressor(
+        n_estimators=n_estimators, random_state=random_state
+    )
+
+    model.fit(X_binned, y)
+    model.predict(X_binned)
+
+    # Extremely Random Forest
+    model = ExtraTreesRegressor(
         n_estimators=n_estimators, random_state=random_state
     )